order-service/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
redis
kafka-python